Chatbot-RAGsMCP_Analysis/MCP_Analysis-main/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
streamlit>=1.22.0
pandas>=1.5.3
polars>=0.20.0
pyarrow>=14.0.1
python-dotenv>=1.0.0
requests>=2.28.2
psycopg2-binary>=2.9.5
PyGithub>=1.58.1
python-dateutil>=2.8.2
argparse>=1.4.0
pandasai>=2.4.2
matplotlib>=3.7.1
Pillow>=10.0.0 